Refactor KakouneWidget
