+++ src/kakounewidget.hpp
@@ -5,14 +5,8 @@
 #include "kakouneclient.hpp"
 #include "kakouneinfobox.hpp"
 #include "kakounemenu.hpp"
+#include "kakounetextedit.hpp"
 #include <QtWidgets>
-
-class TabIgnoreFilter: public QObject
-{
-  Q_OBJECT
-  protected:
-    bool eventFilter(QObject *object, QEvent *event);
-};
 
 class KakouneWidget : public QWidget
 {
@@ -21,17 +15,8 @@ class KakouneWidget : public QWidget
     KakouneWidget(const QString &session_id, DrawOptions *draw_options, QWidget *parent = nullptr);
     ~KakouneWidget();
 
-    void pressKey(int key_code, Qt::KeyboardModifiers modifiers);
-
     KakouneClient *getClient();
 
-  protected:
-    void paintEvent(QPaintEvent *ev) override;
-    void keyPressEvent(QKeyEvent *ev) override;
-    void mouseMoveEvent(QMouseEvent *ev) override;
-    void mousePressEvent(QMouseEvent *ev) override;
-    void mouseReleaseEvent(QMouseEvent *ev) override;
-    void resizeEvent(QResizeEvent *ev) override;
   private slots:
     void clientRefreshed();
   signals:
@@ -41,6 +26,7 @@ class KakouneWidget : public QWidget
   private:
     KakouneClient *m_client;
 
+    KakouneTextEdit *m_textedit;
     KakouneMenu *m_menu;
     KakouneInfoBox *m_info_box;
